7 create issuefeature templates (#11)
This will add templates so when a user clicks the issues / new issue
button it provides an interactive window to guide them on their issue.

I have it setup in the yaml files so that it asks what type of issue for
each script is being reported.

Ideally if we get #9 setup we *should* be able to also provide those
codeowners the ability to edit their specific issue files and also
assign issues automatically to those teams!

## EDIT: I tried to make these where we could have one template per
script class

i.e. CWMS, USGS, etc

But ran into issues with trying to setup conditions in the YML. I.e.
`if` is not a supported value in the github yaml parser.

Redid them so we have a bug and feature issue. Then you select the
script in those.

I also took some time to make a config.yml that points to the other
repos in case their issue is with those and not cwms-cli!