Add CSV2CWMS Script (#16)
Adds a cli script option to read a config + csv file and load it into
the database.
i.e.
```bash
cwms-cli csv2cwms -o swt \
-c "cwmscli\commands\csv2cwms\tests\data\sample_config.json" \
--data-file="cwmscli\commands\csv2cwms\tests\data\sample_brok.csv" \
--begin="2025-03-25T22:15" \
--verbose \
--dry-run
```


Some changes in the logic can be made and also a README to be written to
overview entire script, usage cases, and some of the less obvious
features

Ability to do math in the CSV column headers

i.e.
`column1+column2 -> same-tsid`

## Todos:
1. Connected `cwmscli\commands\csv2cwms\tests` to CI/CD
2. Write documentation (README.md?)
3. Rework logic
  - Update to use begin for start of times (Default to start of file)
  - Update to use end for end of times (default to end of file)
  - Allow optional lookback from the end time
- Attempt to use the interval provided by the timeseries INTERVAL part
and convert that for use
- In the console windows, could we see the oldest datetime in the CSV
file and the newest datetime in the CSV file? And possibly the increment
of data in the file?
  - is -dp required if path to csv is spelled out? Make dp optional
- Add ability to specify a date format in the args i.e. --date-format
https://github.com/HydrologicEngineeringCenter/cwms-cli/issues/3



## Key Notes
- Had to rename directory `cwms` to `commands` because of import
conflicts with `import cwms`
- Converted from using urllib to cwms-python for store-timeseries
- Tested against database