20 setup read the docs documentation (#31)
I tried to set this up so that it could also use sphinx-click to
automatically show the commands from the click arguments we are using.

Part of that and using sphinx was warning me and would not build if I
tried to run python `3.9`.

@Enovotny if you are working on 3.12 or 3.13 for the T7, is it
reasonable to mark 3.11 as the minimum for sphinx/click to work in the
RTD gen? ( was not able to generate with the minimum in the toml set to
3.9 or 3.10 )

I also setup a CI/CD action that should make it so a PR fails if the
docs fail to build. Action should only run if one of those directories
change(?)