Make PyPi test only run when needed (#76)
Previously we would get failed builds because even if the PR was merged
if the version did not change or did not need to change it would fail.

Doing it this way lets us explicitly bump a version in between PRs and
also make doc/tests changes without having to publish a version.

It requires specific directories be edited AND a diff on the last
committed `/pyproject.toml` version to the current

Mike if you wouldn't mind reviewing this action. There maybe a better
way (already made action?) but I really believe we don't need this to
run on every commit.